Draft: Build groups Of grid tasks vectors based on frequent tri-concepts given by TRIAS

+++ src/main/java/de/unikassel/cs/kde/trias/util/TriasCommandLineArgumentsConfigurator.java
@@ -90,7 +90,7 @@ public class TriasCommandLineArgumentsConfigurator implements TriasConfigurator 
 		/*
 		 * minimal supports
 		 */
-		final int[] minSupportPerDimension = new int[Dimension.noOfDimensions];
+		final float[] minSupportPerDimension = new float[Dimension.noOfDimensions];
 		for (int dim = 0; dim < minSupportPerDimension.length; dim++) {
 			minSupportPerDimension[dim] = nextIntArg(); 
 		}
